Release/2.0.0 beta.1 into Master

+++ frontend/src/components/dataset/ResourceCard.vue
@@ -19,18 +19,18 @@
                         <label left class="sublabel">
                             {{useResource.metadata.format}}
                         </label>
-                        <span v-if="useResource.metadata.size" class="sublabel ml-4">
+                        <span v-if="useResource.metadata.size && useResource.metadata.size > 0" class="sublabel ml-4">
                             {{(useResource.metadata.size/1000).toFixed(1)}} MB
                         </span>
                     </span>
                     <span class="floatRight">
                         <v-btn v-if="!loadPOW" class="px-0" small depressed text :href="useResource.metadata.url" color="label_colour">
-                            Download
+                            {{$tc("Access/Download")}}&nbsp;
                         </v-btn>
                         <powButton v-else :btn="true" :resource="useResource.metadata"/>
                         <v-btn v-if="!datasetBeingEdited" small depressed text class="px-0" color="label_colour"
                                 :to="{ name: 'resource_view', params: { datasetId: dataset.name, resourceId: resource.id}}">
-                            View
+                            {{$tc("View")}}
                         </v-btn>
                         <v-btn v-if="!!useResource.hasSchema" small depressed text class="px-0" @click.stop="schemaDialog = true" color="label_colour">
                             View Schema (JSON Table Schema)
@@ -73,9 +73,11 @@ export default {
     },
     methods: {
         deleteResource: function(){
-            ckanServ.deleteResource(this.resource.id).then( () => {
-                location.reload();
-            });
+            if (confirm("Are you sure you want to delete this resource?")) {
+                ckanServ.deleteResource(this.resource.id).then( () => {
+                    location.reload();
+                });
+            }
         }
     },
     computed: {
